# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: cmp-security是一个客户关系管理权限平台
  version: 1.0.0
  title: 客户关系管理权限平台
basePath: /cmp-security
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/role/list:
    get:
      tags:
        - 角色管理
      summary: 角色列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: 角色名
          in: query
          type: string
        - name: roleCode
          description: 角色代码
          in: query
          type: string
      responses:
        '200':
          description: 返回角色列表
          schema:
            $ref: '#/definitions/SysRoleEntityList'
...
  /sys/role/info/{roleId}:
    get:
      tags:
        - 角色管理
      summary: 获取角色信息
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              role:
                $ref: '#/definitions/SysRoleEntity'
  /sys/role/save:
    post:
      tags:
        - 角色管理
      summary: 添加角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/update:
    post:
      tags:
        - 角色管理
      summary: 修改角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/delete:
    post:
      tags:
        - 角色管理
      summary: 删除角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/isValid/{roleId}:
    post:
      tags:
        - 角色管理
      summary: 是否失效
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/queryUserListByroleId:
    get:
      tags:
        - 角色管理
      summary: 通过roleId查询属于该role的用户
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: query
          type: integer
          format: int32
          required: true
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 用户名
          in: query
          type: string
        - name: realname
          description: 真实名
          in: query
          type: string
        - name: startTime
          description: 开始时间
          in: query
          type: date
        - name: endTime
          description: 结束时间
          in: query
          type: date
      responses:
        '200':
          description: 返回用户列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
  /sys/role/notExistCurrRole:
    get:
      tags:
        - 角色管理
      summary: 通过roleId查询不再该角色下的用户
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: query
          type: integer
          format: int32
          required: true
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 用户名
          in: query
          type: string
        - name: realname
          description: 真实名
          in: query
          type: string
        - name: startTime
          description: 开始时间
          in: query
          type: date
        - name: endTime
          description: 结束时间
          in: query
          type: date
      responses:
        '200':
          description: 返回用户列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
  /sys/role/notExistCurrRoleAddUser:
    post:
      tags:
        - 角色管理
      summary: 为不再当前角色下的用户添加该角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit1'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/findRoleByUserId:
    get:
      tags:
        - 角色管理
      summary: 通过用户ID查看他创建的角色
      produces:
        - application/json
      parameters:
        - name: userId
          description: 创建者ID-->本次登录用户的Id
          in: query
          type: integer
          format: int32
          required: true
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: 角色名
          in: query
          type: string
        - name: roleCode
          description: 角色代码
          in: query
          type: string
      responses:
        '200':
          description: 返回用户列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
  /sys/role/findParentNodeByMenuId/{menuId}:
    get:
      tags:
        - 角色管理
      summary: 获取角色信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              role:
                $ref: '#/definitions/SysRoleEntity'
  /sys/role/repeatRoleCode/{roleCode}:
    get:
      tags:
        - 角色管理
      summary: 通过角色代码判断该角色是否存在
      produces:
        - application/json
      parameters:
        - name: roleCode
          description: 角色代码roleCode
          in: path
          type: String
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      realname:
        description: 用户真实名
        type: string
      extension:
        description: 分机号
        type: string
      isExtension:
        description: 是否绑定分机号
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      email:
        description: 邮箱
        type: string
      mobile:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createTime:
        description: 创建时间
        type: string
        format: date
      expirationTime:
        description: 账号过期时间
        type: string
        format: date
      sysDeptEntity:
        type: array
        items:
          $ref: '#/definitions/sysDeptEntity'
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
...
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date
  SysRoleEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysRoleEntity'
  SysRoleEntityEdit:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleCode:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      deptIdList:
        description: 部门ID列表
        type: array
        items:
          type: integer
          format: int64
      creater:
        description: 用户ID
        type: integer
        format: int64
  SysRoleEntityEdit1:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      userIdList:
        description: 用户ID列表
        type: array
        items:
          type: integer
          format: int64
...